@@ -4,14 +4,10 @@ import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { BookListComponent } from './book-list.component';
-import { BookDetailComponent } from './book-detail.component';
-import { routing } from './app.routing';
 
 @NgModule({
-  declarations: [AppComponent, BookDetailComponent, BookListComponent],
-  imports: [BrowserModule, FormsModule, HttpClientModule, routing],
+  declarations: [AppComponent],
+  imports: [BrowserModule, FormsModule, HttpClientModule],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
-
