@@ -7,13 +7,13 @@ import { HttpClient } from '@angular/common/http';
 })
 export class AppComponent {
 
-  books: string[] = [];
+  titles: string[] = [];
 
   constructor(private httpClient: HttpClient) { }
 
   search(title: string) {
 
-    this.books = [];
+    this.titles = [];
 
     let url = "https://www.googleapis.com/books/v1/volumes?q=intitle:" + title;
 
@@ -22,7 +22,7 @@ export class AppComponent {
         let data: any = response;
         for (var i = 0; i < data.items.length; i++) {
           let bookTitle = data.items[i].volumeInfo.title;
-          this.books.push(bookTitle);
+          this.titles.push(bookTitle);
         }
       },
       error => console.error(error)
