@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 @Injectable({ providedIn: 'root' })
 export class BooksService {
 
-  getBooks(title: string) {
+  getTitles(title: string) {
     return [
       'Aprende Java en 2 días',
       'Java para torpes',
