@@ -1,10 +1,18 @@
 import { Component } from '@angular/core';
+import { BooksService } from './books.service';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-root',
-  template: `
-    <h1 class="title">Library</h1>
-    <router-outlet></router-outlet>
-  `
+  templateUrl: './app.component.html'
 })
-export class AppComponent { }
+export class AppComponent {
+
+  $titles: Observable<string>;
+
+  constructor(private service: BooksService) { }
+
+  search(title: string) {
+    this.$titles = this.service.getTitles(title);
+  }
+}
