@@ -1,37 +1,12 @@
 import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { MaterialModule } from './material.module';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { WelcomeComponent } from './welcome/welcome.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { AppRouters } from './app.routes';
-import { DataService } from './data/data.service';
-import { AuthService } from './auth.service';
-import { PostDialogComponent } from './post-dialog/post-dialog.component';
-import { FormsModule } from '@angular/forms';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    WelcomeComponent,
-    DashboardComponent,
-    PostDialogComponent
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    MaterialModule,
-    FlexLayoutModule,
-    AppRouters,
-    FormsModule,
-  ],
-  providers: [DataService, AuthService],
-  bootstrap: [AppComponent],
-  entryComponents: [
-    PostDialogComponent
-  ]
+  declarations: [AppComponent],
+  imports: [BrowserModule, FormsModule],
+  bootstrap: [AppComponent]
 })
 export class AppModule { }
