@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
+
 import { NavController } from 'ionic-angular';
-import { DetailsPage } from '../details/details';
 
 @Component({
   selector: 'page-home',
@@ -8,16 +8,8 @@ import { DetailsPage } from '../details/details';
 })
 export class HomePage {
 
-  public elems = [
-      {name:"Elem1", desc: "Elem1 description"},
-      {name:"Elem2", desc: "Elem2 description"},
-      {name:"Elem3", desc: "Elem3 description"}
-    ];
+  constructor(public navCtrl: NavController) {
 
-    constructor(private navCtrl: NavController) {}
+  }
 
-    goToDetails(elem){
-      this.navCtrl.push(DetailsPage, {elem: elem})
-    }
 }
-
