@@ -1,29 +1,15 @@
 import { Component } from '@angular/core';
-import { NavParams, NavController } from 'ionic-angular';
-import { GitHubService } from '../../services/github';
+
+import { NavParams} from 'ionic-angular';
 
 @Component({
     templateUrl: 'details.html'
 })
 export class DetailsPage {
 
-    public readme = '';
-    public repo;
+    elem: any;
 
-    constructor(private github: GitHubService,
-                private nav: NavController, private navParams: NavParams) {
-
-        this.repo = navParams.get('repo');
-        this.github.getDetails(this.repo).subscribe(
-            data => this.readme = data.text(),
-            err => {
-                if (err.status == 404) {
-                    this.readme = 'This repo does not have a README. :(';
-                } else { 
-					console.error(err);
-				}
-            },
-            () => console.log('getDetails completed')
-        );
+    constructor(navParams: NavParams) {
+        this.elem = navParams.get('elem');
     }
 }
